requirements.txt
streamlit
geopandas
pyarrow
duckdb
pydeck
altair
pygwalker